refactor(web): align shared layout and session ports
