+++ web/src/lib/components/chat-board/ChatBoardLane.svelte
@@ -6,7 +6,7 @@
 		ChatBoardLaneProjection,
 		ChatBoardOccurrence,
 	} from '$lib/chat-board/projection/chat-board-projection.js';
-	import type { ChatItemLayout } from '$lib/chat/presentation/chat-item-layout.js';
+	import type { ChatItemLayout } from '$lib/layout/chat-item-layout.js';
 	import * as m from '$lib/paraglide/messages.js';
 	import { CHAT_BOARD_COLUMN_DROP_TYPE, isChatBoardCardDragData } from './chat-board-dnd.js';
 	import ChatBoardLaneList from './ChatBoardLaneList.svelte';
