+++ web/src/lib/components/sidebar/SidebarControlsRow.svelte
@@ -30,7 +30,7 @@
 		SidebarChatGrouping,
 		SidebarSortMode,
 	} from '$lib/stores/local-settings.svelte';
-	import type { ChatItemLayout } from '$lib/chat/presentation/chat-item-layout.js';
+	import type { ChatItemLayout } from '$lib/layout/chat-item-layout.js';
 	import type { SavedChatSearch } from '$lib/api/settings';
 	import { sidebarGroupingUsesProjects } from './sidebar-display-options';
