+++ web/src/lib/components/sidebar/SidebarVirtualChatListController.svelte.ts
@@ -1,5 +1,5 @@
 import type { Attachment } from 'svelte/attachments';
-import type { ChatItemLayout } from '$lib/chat/presentation/chat-item-layout.js';
+import type { ChatItemLayout } from '$lib/layout/chat-item-layout.js';
 import { VirtualListController } from '$lib/virt/virtual-list-controller.svelte.js';
 import {
 	virtualItems as selectVirtualItems,
