+++ web/src/lib/components/sidebar/sidebar-display-options.ts
@@ -3,7 +3,7 @@ import type {
 	SidebarInactivityDuration,
 	SidebarSortMode,
 } from '$lib/stores/local-settings.svelte';
-import type { ChatItemLayout } from '$lib/chat/presentation/chat-item-layout.js';
+import type { ChatItemLayout } from '$lib/layout/chat-item-layout.js';
 import type { PinnedInsertPosition } from '$shared/settings';
 
 export interface SidebarDisplayOptions {
